+++ src/sharedHelpers/persistedInstallationId.ts
@@ -7,15 +7,15 @@ const INSTALL_ID = "installID";
 // one wrapping zustand because we want to be able to persist the installation ID
 // even if a user logs out of the app. this means on signout, we're clearing
 // the storage for the MMKV instance wrapping zustand but leaving this one untouched
-const userData = new MMKV( {
+const persistedInstallationId = new MMKV( {
   id: "user-data"
 } );
-export default userData;
+export default persistedInstallationId;
 
 export function installID( ) {
-  const id = userData.getString( INSTALL_ID );
+  const id = persistedInstallationId.getString( INSTALL_ID );
   if ( id ) return id;
   const newID: string = uuid.v4( ).toString( );
-  userData.set( INSTALL_ID, newID );
+  persistedInstallationId.set( INSTALL_ID, newID );
   return newID;
 }
